djangoProject/requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
